requirements.txt
altgraph==0.17.4
bleak==0.22.3
data==0.4
decorator==5.2.1
funcsigs==1.0.2
macholib==1.16.3
numpy==2.2.5
packaging==25.0
paho-mqtt==2.1.0
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-CoreBluetooth==10.3.2
pyobjc-framework-libdispatch==10.3.2
pyqtgraph==0.13.7
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
qasync==0.27.1
setuptools==80.9.0
shiboken6==6.9.0
six==1.17.0
